// frontend/src/types/schema.d.ts
export interface components {
    schemas: {
        "handlers.CreateCommentRequest": {
            content: string;
        };
        "handlers.CreatePostRequest": {
            content: string;
        };
        "handlers.LoginRequest": {
            email: string;
            password: string;
        };
        "handlers.RegisterRequest": {
            email: string;
            password: string;
            username: string;
        };
        "handlers.UpdatePostRequest": {
            content: string;
        };
        "handlers.UpdateProfileRequest": {
            avatar_url?: string;
            bio?: string;
            birth_date?: string;
            display_name?: string;
            header_url?: string;
            occupation?: string;
            website?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}